Parse stderr services from "service --status-all"
This requires some underlying architecture changes. We now need to be
able to merge together multiple services. Since the zodb recorder has to
support merging, now all the other parsers *also* need to support
merging. So convert from using a hash as the list of
services/paths/whatever, to using a subclass of
collections.MutableMapping